Cleanup and fixes

+++ pyrxiv/cli/cli.py
@@ -1,16 +1,12 @@
 import re
 import time
 from pathlib import Path
-from typing import TYPE_CHECKING
 
 import click
 import h5py
 import numpy as np
 
-if TYPE_CHECKING:
-    from pyrxiv.datamodel import ArxivPaper
-
-
+from pyrxiv.datamodel import ArxivPaper
 from pyrxiv.download import ArxivDownloader
 from pyrxiv.extract import TextExtractor
 from pyrxiv.fetch import ArxivFetcher
@@ -46,19 +42,21 @@ def run_search_and_download(
     start_from_filepath: bool = False,
     loader: str = "pdfminer",
     clean_text: bool = True,
+    download_pdfs: bool = False,
 ) -> tuple[list[Path], list["ArxivPaper"]]:
     """
     Searches for a specific number of papers `n_papers` in arXiv for a specified `category` and downloads
-    them in a `download_path`.
+    their metadata in an HDF5 file in `download_path`.
 
     If `regex_pattern` is specified, only the papers that contain the pattern will be downloaded.
     If `start_id` is specified, the search will start from that ID.
     If `start_from_filepath` is True, the search will start from the last downloaded paper's ID.
     If `loader` is specified, the text will be extracted using the corresponding loader.
-
+    If `clean_text` is True, the extracted text will be cleaned by removing references and unnecessary whitespaces.
+    If `download_pdfs` is True, the PDFs will be downloaded and saved in `download_path`.
 
     Args:
-        download_path (Path, optional): The path for downloading the arXiv PDFs. Defaults to Path("data").
+        download_path (Path, optional): The path for downloading the arXiv metadata. Defaults to Path("data").
         category (str, optional): The arXiv category on which the papers will be searched. Defaults to "cond-mat.str-el".
         n_papers (int, optional): The number of arXiv papers to be fetched and downloaded.
             If `regex_pattern` is not specified, this would correspond to the n_papers starting from the newest in the `category`. Defaults to 5.
@@ -70,9 +68,11 @@ def run_search_and_download(
         loader (str, optional): PDF loader to use for extracting text from the downloaded PDFs.
             Defaults to "pdfminer". Available loaders: "pdfminer", "pypdf".
         clean_text (bool, optional): If True, the extracted text will be cleaned by removing references and unnecessary whitespaces.
+            Defaults to True.
+        download_pdfs (bool, optional): If True, the PDFs will be downloaded and saved in `download_path`. Defaults to False.
 
     Returns:
-        tuple[list[Path], list[ArxivPaper]]: A tuple containing a list of Paths to the downloaded PDFs and a list of ArxivPaper objects
+        tuple[list[Path], list[ArxivPaper]]: A tuple containing a list of Paths to the arXiv papers and a list of ArxivPaper objects
             with the extracted text.
     """
     if loader not in ["pdfminer", "pypdf"]:
@@ -134,7 +134,8 @@ def run_search_and_download(
                 save_paper_to_hdf5(paper=paper, pdf_path=pdf_path, hdf_path=hdf_path)
 
                 # Deleting the PDF file after storing it in HDF5
-                pdf_path.unlink()
+                if not download_pdfs:
+                    pdf_path.unlink()
 
                 # Appending the HDF5 file and paper to the lists
                 pattern_files.append(hdf_path)
@@ -162,7 +163,7 @@ def cli():
     default="data",
     required=False,
     help="""
-    (Optional) The path for downloading the arXiv PDFs. Defaults to "data".
+    (Optional) The path for downloading the arXiv metadata in HDF5 files and, optionally (if set with download-pdfs), the PDFs. Defaults to "data".
     """,
 )
 @click.option(
@@ -239,6 +240,16 @@ def cli():
     Defaults to True.
     """,
 )
+@click.option(
+    "--download-pdfs",
+    "-dp",
+    is_flag=True,
+    default=False,
+    required=False,
+    help="""
+    (Optional) If True, the PDFs will be downloaded and saved in `download_path`. Defaults to False.
+    """,
+)
 def search_and_download(
     download_path,
     category,
@@ -248,6 +259,7 @@ def search_and_download(
     start_from_filepath,
     loader,
     clean_text,
+    download_pdfs,
 ):
     start_time = time.time()
 
@@ -260,7 +272,56 @@ def search_and_download(
         start_from_filepath=start_from_filepath,
         loader=loader,
         clean_text=clean_text,
+        download_pdfs=download_pdfs,
     )
 
     elapsed_time = time.time() - start_time
     click.echo(f"Downloaded arXiv papers in {elapsed_time:.2f} seconds\n\n")
+
+
+@cli.command(
+    name="download_pdfs",
+    help="Downloads the PDFs of the arXiv papers stored in HDF5 files in a specified path.",
+)
+@click.option(
+    "--data-path",
+    "-path",
+    type=str,
+    default="data",
+    required=False,
+    help="""
+    (Optional) The path where the HDF5 files with the arXiv papers metadata exist. The downloaded PDFs will be stored in there as well. Defaults to "data".
+    """,
+)
+def download_pdfs(data_path):
+    start_time = time.time()
+
+    # check if `data_path` exists, and if not, returns an error
+    data_path = Path(data_path)
+    if not data_path.exists():
+        raise click.ClickException(f"The specified path {data_path} does not exist.")
+    downloader = ArxivDownloader(download_path=data_path, logger=logger)
+
+    # Loops over all HDF5 files in the `data_path` and downloads the corresponding PDFs
+    hdf5_files = list(data_path.glob("*.hdf5"))
+
+    failed_downloads = []
+    with click.progressbar(
+        length=len(hdf5_files), label="Downloading papers PDFs"
+    ) as bar:
+        for file in hdf5_files:
+            paper = ArxivPaper.from_hdf5(file=file)
+            try:
+                _ = downloader.download_pdf(arxiv_paper=paper)
+            except Exception as e:
+                failed_downloads.append(str(file))
+                logger.error(f"Failed to download PDF for {file}: {e}")
+            bar.update(1)
+
+    elapsed_time = time.time() - start_time
+    click.echo(f"Downloaded arXiv papers in {elapsed_time:.2f} seconds\n\n")
+
+    if failed_downloads:
+        click.echo("\nFailed to download PDFs for the following files:")
+        for failed_file in failed_downloads:
+            click.echo(f"  - {failed_file}")
